+++ ChangeLog.md
@@ -1,3 +1,9 @@
+# v0.3.0 - 02/7/2023
+
+- Major code cleanup and documentation
+- Removal of recurrnt events
+- APIs `pop` and `wait` can now be freely mixed
+
 # v0.2.0 - 31/5/2023
 
 - Major API cleanup and documentation